renamed to handleBrowserForestManagerUpdates
