+++ packages/pwa-builder-manager/src/handleBrowserForestManagerUpdates.ts
