+++ packages/pwa-builder-manager/src/index.ts
@@ -2,4 +2,4 @@ export * from "./init";
 export * from "./api";
 export * from "./canvasApi";
 export * from "./breakpointApi";
-import "./handleBrowserForestManagerEvents";
+import "./handleBrowserForestManagerUpdates";
